requirements.txt
streamlit
pymupdf
requests
google-generativeai
pytesseract
Pillow
pdf2image
gTTS
SpeechRecognition
streamlit-audiorecorder